Doc update: Correct magic values from nbd protocol, V2
The magic number documentation refers to old values for
NBD_REQUEST_MAGIC and NBD_REPLY_MAGIC: The documented values were used
until Linux 2.1.116pre2.

Thus:
- Update the documentation.
- Update the header file: The authorative source for the nbd protocol
  is proto.md from the nbd package, thus mention this.
- Remove the historic values from the header file.
  The historic values are still documented in proto.md from the nbd
  package.

Removing the historic values is intentional:
The values are stale for > 20 years, and this was not noticed.
My guess is that everyone used grep to confirm that the values are
still in use - and the historic values resulted that there were
still hits with grep, ...

Signed-off-by: Manfred Spraul <manfred.spraul@de.bosch.com>
Link: https://github.com/NetworkBlockDevice/nbd/commit/107356ee528eb30744d518a8ac1cb6d379da4868
Link: https://lore.kernel.org/all/20220318200446.14648-1-manfred@colorfullife.com/
Link: https://lists.debian.org/nbd/2022/01/msg00039.html
Link: https://lore.kernel.org/r/20220805082532.55131-1-manfred@colorfullife.com
Signed-off-by: Jonathan Corbet <corbet@lwn.net>